Integrates the parallel main-line work with the audit-gap remediation branch:
- v0.11.0 version alignment across charts/contracts/services/docs,
- paper/ ruff fixes + its registration in the canonical repository layout,
- the gateway /readyz public-probe auth exemption and the runtime_client "/"
  ollama health-check fallback (with their regression tests).

Conflicts resolved in favour of the branch where they were pure version strings
(v0.12.0 supersedes the v0.11.0 strings). main's surgical gateway fixes were
preserved (the /readyz exemption in _auth_required, the /readyz removal from the
api-contract protected_paths, and the runtime_client "/" fallback), the OpenAPI
snapshots and chart docs were regenerated from the resolved sources, and paper/
is kept in lint scope -- its pyproject exclude was dropped now that main has made
it ruff-clean, and the duplicate paths.py "paper" entry was collapsed to one.

make validate passes end-to-end; gateway 115 tests (including main's two /readyz
regression tests) and rag 51 tests pass.

+++ paper/cost-of-compliance/bench.py
@@ -38,7 +38,7 @@ def _percentile(values: list[float], pct: float) -> float:
         return ordered[0]
     if pct >= 100:
         return ordered[-1]
-    rank = max(1, int(round(pct / 100.0 * len(ordered))))
+    rank = max(1, round(pct / 100.0 * len(ordered)))
     return ordered[min(rank, len(ordered)) - 1]
